+++ flake.nix
@@ -10,8 +10,11 @@
     nixpkgs-lock.url = "github:pr0d1r2/nixpkgs-lock";
     nixpkgs.follows = "nixpkgs-lock/nixpkgs";
 
-    set-and-setting.url = "github:pr0d1r2/set-and-setting";
-    set-and-setting.inputs.nixpkgs-lock.follows = "nixpkgs-lock";
+    set-and-setting = {
+      url = "github:pr0d1r2/set-and-setting";
+      inputs.nixpkgs.follows = "nixpkgs";
+      inputs.nixpkgs-lock.follows = "nixpkgs-lock";
+    };
   };
 
   outputs =
@@ -23,16 +26,6 @@
     }:
     set-and-setting.lib.mkConsumerFlake {
       inherit self nixpkgs set-and-setting;
-      lib = set-and-setting.lib // {
-        checksFor =
-          args:
-          set-and-setting.lib.checksFor (
-            args
-            // {
-              fragments = builtins.filter (fragment: fragment != "actions") args.fragments;
-            }
-          );
-      };
       fragments = [
         "base"
         "actions"